+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## v0.5.0
 - [Major] Upgrade to htmlparser2 module for faster parsing and more options.
+- [Major] Mimimize as constructor to allow multiple instances without infecting.
+- [Minor] Remove CDATA from scripts by default
 - [Fix] Use previous reference of element for checking against text
 - [Tests] Add some tests against above mentioned functionality